python-api/requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
requests==2.32.3